+++ packages/pixi-solid/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.3
 
-- Improved error handling for invalid children.
+- ✅ Improved error handling for invalid children.
 
 ## 0.1.2